Get Information on Line Damages

Get Information on Line Damages

getLineDamageInfo(
  observations,
  I = observations$I,
  J = observations$J,
  inspno = observations$inspno,
  dbg = TRUE
)

Arguments

observations

data frame with observations. Required columns: I (= horizontal or vertical position), J (= code for line damage), inspno (number of inspection to which the observation belongs)

I

Numeric vector of positions of observations

J

Vector of character of same length as I containing the codes used to indicate continuous defects

inspno

vector of inspection identifiers of same length as I

dbg

if TRUE, debug messages are shown

Value

data frame with columns ino (inspection number), ldno (line damage number), beg.at, end.at, beg.x (position of line damage begin), end.x (position of line damage end), length (length of line damage)