Plot Observations per Pipe

Plot Observations per Pipe

plotObservations(survey, to_pdf = TRUE, matrix_dim = c(7, 1))

Arguments

survey

list with elements inspections and observations as e.g. returned by kwb.en13508.2:::readEuCodedFile

to_pdf

if TRUE (default) the output goes into a temporary PDF file

matrix_dim

vector of two specifying the number of rows and columns, respectivly of the matrix in which to arrange the single plots.

Examples

# Install packages if not yet installed
if (FALSE) {
install.packages("ggplot2")
devtools::install_github("guiastrennec/ggplus")
}

# Get example data that are contained in the package
survey <- kwb.en13508.2:::getExampleData()
#> Error in stopOnInvalidEncoding(file.encoding): No such encoding string:
#> - 'latin1'
#> Available encoding strings (sorted):
#> - '437'
#> - '850'
#> - '852'
#> - '855'
#> - '857'
#> - '860'
#> - '861'
#> - '862'
#> - '863'
#> - '865'
#> - '866'
#> - '869'
#> - 'ANSI_X3.4-1968'
#> - 'ANSI_X3.4-1986'
#> - 'ASCII'
#> - 'ASMO-708'
#> - 'BIG-5'
#> - 'BIG-FIVE'
#> - 'BIG5'
#> - 'BIG5-HKSCS'
#> - 'BIG5HKSCS'
#> - 'CP-GR'
#> - 'CP-IS'
#> - 'CP1125'
#> - 'CP1133'
#> - 'CP1200'
#> - 'CP12000'
#> - 'CP12001'
#> - 'CP1201'
#> - 'CP1250'
#> - 'CP1251'
#> - 'CP1252'
#> - 'CP1253'
#> - 'CP1254'
#> - 'CP1255'
#> - 'CP1256'
#> - 'CP1257'
#> - 'CP1258'
#> - 'CP1361'
#> - 'CP154'
#> - 'CP367'
#> - 'CP437'
#> - 'CP50221'
#> - 'CP51932'
#> - 'CP65001'
#> - 'CP737'
#> - 'CP775'
#> - 'CP819'
#> - 'CP850'
#> - 'CP852'
#> - 'CP853'
#> - 'CP855'
#> - 'CP857'
#> - 'CP858'
#> - 'CP860'
#> - 'CP861'
#> - 'CP862'
#> - 'CP863'
#> - 'CP864'
#> - 'CP865'
#> - 'CP866'
#> - 'CP869'
#> - 'CP874'
#> - 'CP932'
#> - 'CP936'
#> - 'CP949'
#> - 'CP950'
#> - 'CSASCII'
#> - 'CSIBM855'
#> - 'CSIBM857'
#> - 'CSIBM860'
#> - 'CSIBM861'
#> - 'CSIBM863'
#> - 'CSIBM864'
#> - 'CSIBM865'
#> - 'CSIBM866'
#> - 'CSIBM869'
#> - 'CSISOLATIN1'
#> - 'CSPC775BALTIC'
#> - 'CSPC850MULTILINGUAL'
#> - 'CSPC862LATINHEBREW'
#> - 'CSPC8CODEPAGE437'
#> - 'CSPCP852'
#> - 'CSPTCP154'
#> - 'CSWINDOWS31J'
#> - 'CYRILLIC-ASIAN'
#> - 'DOS-720'
#> - 'DOS-862'
#> - 'EUC-CN'
#> - 'EUC-KR'
#> - 'EUCCN'
#> - 'GB18030'
#> - 'GBK'
#> - 'IBM-CP1133'
#> - 'IBM-Thai'
#> - 'IBM00858'
#> - 'IBM00924'
#> - 'IBM01047'
#> - 'IBM01140'
#> - 'IBM01141'
#> - 'IBM01142'
#> - 'IBM01143'
#> - 'IBM01144'
#> - 'IBM01145'
#> - 'IBM01146'
#> - 'IBM01147'
#> - 'IBM01148'
#> - 'IBM01149'
#> - 'IBM037'
#> - 'IBM1026'
#> - 'IBM273'
#> - 'IBM277'
#> - 'IBM278'
#> - 'IBM280'
#> - 'IBM284'
#> - 'IBM285'
#> - 'IBM290'
#> - 'IBM297'
#> - 'IBM367'
#> - 'IBM420'
#> - 'IBM423'
#> - 'IBM424'
#> - 'IBM437'
#> - 'IBM437'
#> - 'IBM500'
#> - 'IBM775'
#> - 'IBM819'
#> - 'IBM850'
#> - 'IBM852'
#> - 'IBM855'
#> - 'IBM855'
#> - 'IBM857'
#> - 'IBM860'
#> - 'IBM860'
#> - 'IBM861'
#> - 'IBM862'
#> - 'IBM863'
#> - 'IBM863'
#> - 'IBM864'
#> - 'IBM864'
#> - 'IBM865'
#> - 'IBM865'
#> - 'IBM866'
#> - 'IBM869'
#> - 'IBM870'
#> - 'IBM871'
#> - 'IBM880'
#> - 'IBM905'
#> - 'ISO-2022-JP'
#> - 'ISO-2022-JP-MS'
#> - 'ISO-8859-1'
#> - 'ISO-IR-100'
#> - 'ISO-IR-6'
#> - 'ISO2022-JP'
#> - 'ISO2022-JP-MS'
#> - 'ISO646-US'
#> - 'ISO8859-1'
#> - 'ISO_646.IRV:1991'
#> - 'ISO_8859-1'
#> - 'ISO_8859-1:1987'
#> - 'JOHAB'
#> - 'Johab'
#> - 'L1'
#> - 'LATIN1'
#> - 'MS-ANSI'
#> - 'MS-ARAB'
#> - 'MS-CYRL'
#> - 'MS-EE'
#> - 'MS-GREEK'
#> - 'MS-HEBR'
#> - 'MS-TURK'
#> - 'MS50221'
#> - 'MS51932'
#> - 'MS932'
#> - 'MS936'
#> - 'PT154'
#> - 'PTCP154'
#> - 'SHIFFT_JIS'
#> - 'SHIFFT_JIS-MS'
#> - 'SJIS'
#> - 'SJIS-MS'
#> - 'SJIS-OPEN'
#> - 'SJIS-WIN'
#> - 'UCS-2'
#> - 'UCS-2BE'
#> - 'UCS-2LE'
#> - 'UCS-4'
#> - 'UCS-4BE'
#> - 'UCS-4BE'
#> - 'UCS-4LE'
#> - 'UCS-4LE'
#> - 'UCS2'
#> - 'UCS2BE'
#> - 'UCS2LE'
#> - 'UCS4'
#> - 'UCS4BE'
#> - 'UCS4LE'
#> - 'UHC'
#> - 'US'
#> - 'US-ASCII'
#> - 'UTF-16'
#> - 'UTF-16BE'
#> - 'UTF-16LE'
#> - 'UTF-32'
#> - 'UTF-32BE'
#> - 'UTF-32LE'
#> - 'UTF-8'
#> - 'UTF16'
#> - 'UTF16BE'
#> - 'UTF16LE'
#> - 'UTF32'
#> - 'UTF32BE'
#> - 'UTF32LE'
#> - 'UTF8'
#> - 'WINBALTRIM'
#> - 'WINDOWS-31J'
#> - 'WINDOWS-50221'
#> - 'WINDOWS-51932'
#> - 'WINDOWS-932'
#> - 'WINDOWS-936'
#> - 'big5'
#> - 'big5-hkscs'
#> - 'big5hkscs'
#> - 'cp1025'
#> - 'cp866'
#> - 'cp875'
#> - 'csISO2022JP'
#> - 'euc-jp'
#> - 'euc-kr'
#> - 'eucjp'
#> - 'euckr'
#> - 'gb2312'
#> - 'hz-gb-2312'
#> - 'ibm737'
#> - 'ibm775'
#> - 'ibm850'
#> - 'ibm852'
#> - 'ibm857'
#> - 'ibm861'
#> - 'ibm869'
#> - 'iso-2022-jp'
#> - 'iso-2022-jp'
#> - 'iso-2022-kr'
#> - 'iso-8859-13'
#> - 'iso-8859-15'
#> - 'iso-8859-2'
#> - 'iso-8859-3'
#> - 'iso-8859-4'
#> - 'iso-8859-5'
#> - 'iso-8859-6'
#> - 'iso-8859-7'
#> - 'iso-8859-8'
#> - 'iso-8859-8-i'
#> - 'iso-8859-9'
#> - 'iso2022-kr'
#> - 'iso8859-1'
#> - 'iso8859-13'
#> - 'iso8859-15'
#> - 'iso8859-2'
#> - 'iso8859-3'
#> - 'iso8859-4'
#> - 'iso8859-5'
#> - 'iso8859-6'
#> - 'iso8859-7'
#> - 'iso8859-8'
#> - 'iso8859-8-i'
#> - 'iso8859-9'
#> - 'iso_8859-1'
#> - 'iso_8859-13'
#> - 'iso_8859-15'
#> - 'iso_8859-2'
#> - 'iso_8859-3'
#> - 'iso_8859-4'
#> - 'iso_8859-5'
#> - 'iso_8859-6'
#> - 'iso_8859-7'
#> - 'iso_8859-8'
#> - 'iso_8859-8-i'
#> - 'iso_8859-9'
#> - 'iso_8859_1'
#> - 'iso_8859_13'
#> - 'iso_8859_15'
#> - 'iso_8859_2'
#> - 'iso_8859_3'
#> - 'iso_8859_4'
#> - 'iso_8859_5'
#> - 'iso_8859_6'
#> - 'iso_8859_7'
#> - 'iso_8859_8'
#> - 'iso_8859_8-i'
#> - 'iso_8859_9'
#> - 'koi8-r'
#> - 'koi8-u'
#> - 'ks_c_5601-1987'
#> - 'latin-9'
#> - 'latin2'
#> - 'latin3'
#> - 'latin4'
#> - 'latin5'
#> - 'latin7'
#> - 'latin9'
#> - 'mac'
#> - 'mac-centraleurope'
#> - 'mac-is'
#> - 'macarabic'
#> - 'maccentraleurope'
#> - 'maccroatian'
#> - 'maccyrillic'
#> - 'macgreek'
#> - 'machebrew'
#> - 'maciceland'
#> - 'macintosh'
#> - 'macis'
#> - 'macroman'
#> - 'macromania'
#> - 'macthai'
#> - 'macturkish'
#> - 'macukraine'
#> - 'macukrainian'
#> - 'shift-jis'
#> - 'shift_jis'
#> - 'unicodeFFFE'
#> - 'windows-1250'
#> - 'windows-1251'
#> - 'windows-1252'
#> - 'windows-1253'
#> - 'windows-1254'
#> - 'windows-1255'
#> - 'windows-1256'
#> - 'windows-1257'
#> - 'windows-1258'
#> - 'windows-874'
#> - 'x-Chinese_CNS'
#> - 'x-EBCDIC-KoreanExtended'
#> - 'x-Europa'
#> - 'x-IA5'
#> - 'x-IA5-German'
#> - 'x-IA5-Norwegian'
#> - 'x-IA5-Swedish'
#> - 'x-cp20001'
#> - 'x-cp20003'
#> - 'x-cp20004'
#> - 'x-cp20005'
#> - 'x-cp20261'
#> - 'x-cp20269'
#> - 'x-cp20936'
#> - 'x-cp20949'
#> - 'x-cp50227'
#> - 'x-iscii-as'
#> - 'x-iscii-be'
#> - 'x-iscii-de'
#> - 'x-iscii-gu'
#> - 'x-iscii-ka'
#> - 'x-iscii-ma'
#> - 'x-iscii-or'
#> - 'x-iscii-pa'
#> - 'x-iscii-ta'
#> - 'x-iscii-te'
#> - 'x-mac-arabic'
#> - 'x-mac-ce'
#> - 'x-mac-chinesesimp'
#> - 'x-mac-chinesetrad'
#> - 'x-mac-croatian'
#> - 'x-mac-cyrillic'
#> - 'x-mac-greek'
#> - 'x-mac-hebrew'
#> - 'x-mac-icelandic'
#> - 'x-mac-japanese'
#> - 'x-mac-korean'
#> - 'x-mac-romanian'
#> - 'x-mac-thai'
#> - 'x-mac-turkish'
#> - 'x-mac-ukrainian'
#> - 'x_Chinese-Eten'

# Create one plot per inspection in "survey"
kwb.en13508.2::plotObservations(survey, to_pdf = FALSE)
#> Error in eval(expr, envir, enclos): object 'survey' not found