Read and Merge Files in EN13508.2-Format

Read files in EN13508.2-format using readEuCodedFiles and merge them by means of mergeInspectionData

readAndMergeEuCodedFiles(
  input.files,
  dbg = FALSE,
  name.convention = "norm",
  ...,
  add.inspid = FALSE,
  project = NULL,
  default.time = "22:22",
  error.file = NULL
)

Arguments

input.files

full path to text file containing CCTV inspection results in the format described in DIN EN 13508-2

dbg

if TRUE debug messages are shown

name.convention

passed to readEuCodedFiles

...

further arguments passed to readEuCodedFiles

add.inspid

if TRUE (the default is FALSE) a globally unique inspection ID (inspid) is added to the data frames in elements "inspections" and "observations" of the returned list.

project

name of project to which the data are related, such as: "Lausanne"

default.time

passed to setGlobalInspectionID

error.file

optional. Path to error file, passed to setGlobalInspectionID.