Write Inspection Data to Files in EU Format

Write inspection data to files in EU format with each file containing data of a fix number (default: 100) of inspections

writeEuCodedFiles(survey, file, blocksize = 100, dbg = TRUE)

Arguments

survey
list with elements header.info, inspections, observations, just as required for parameter "inspection.data" in writeEuCodedFile.

file
Full path to output file. The file name must end with ".txt" which will be replaced with "_<i>_<j>.txt" with i and j being the number of the first and last inspection, respectively, contained in the file.

blocksize
number of inspections to be written to one file. Default: 100

dbg
if TRUE (default) debug messages are shown.

See also